# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "0.4.0",
    "description": "Contract for the current Next.js HTTP routes. Version follows the repository release line. This project is pre-1.0: request and response details may change between releases. See api-contract.md for supported guarantees and known gaps.",
    "title": "Nuera Quicksilver HTTP API",
    "summary": "Pre-1.0 API contract; not a stable compatibility promise."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agents/drafts": {
      "post": {
        "operationId": "createAgentDraft",
        "tags": [
          "Agents"
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/AgentDefinitionResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateAgentDraft"
        }
      }
    },
    "/api/agents/drafts/submit": {
      "post": {
        "operationId": "submitAgentDraft",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/AgentDefinitionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentVersionAction"
        }
      }
    },
    "/api/agents/publish": {
      "post": {
        "operationId": "publishAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/AgentDefinitionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/AgentVersionAction"
        }
      }
    },
    "/api/agents/review": {
      "post": {
        "operationId": "reviewAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/AgentDefinitionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/ReviewAgentDefinition"
        }
      }
    },
    "/api/agents/rollback": {
      "post": {
        "operationId": "rollbackAgent",
        "tags": [
          "Agents"
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/AgentDefinitionResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/RollbackAgentDefinition"
        }
      }
    },
...
    "/api/workflows/drafts": {
      "post": {
        "operationId": "saveWorkflowDraft",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateWorkflowDraftRequest"
        }
      }
    },
    "/api/workflows/drafts/submit": {
      "post": {
        "operationId": "submitWorkflowDraft",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/WorkflowVersionAction"
        }
      }
    },
...
    "/api/workflows/publish": {
      "post": {
        "operationId": "publishWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/WorkflowVersionAction"
        }
      }
    },
    "/api/workflows/review": {
      "post": {
        "operationId": "reviewWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/ReviewWorkflowRequest"
        }
      }
    },
    "/api/workflows/rollback": {
      "post": {
        "operationId": "rollbackWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/WorkflowVersionAction"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkflowDraftRequest": {
        "type": "object",
        "required": ["graph"],
        "additionalProperties": false,
        "properties": { "graph": { "$ref": "#/components/schemas/WorkflowGraph" } }
      },
      "WorkflowVersionAction": {
        "type": "object",
        "required": ["workflowId", "version"],
        "additionalProperties": false,
        "properties": { "workflowId": { "type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]{0,127}$" }, "version": { "type": "integer", "minimum": 1 } }
      },
      "ReviewWorkflowRequest": {
        "type": "object",
        "required": ["workflowId", "version", "note"],
        "additionalProperties": false,
        "properties": { "workflowId": { "type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:-]{0,127}$" }, "version": { "type": "integer", "minimum": 1 }, "note": { "type": "string", "minLength": 10, "maxLength": 500 } }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code. Always present on error responses from the decision, workflow, agent and company-data routes. The value follows the status: invalid-request (400), unauthenticated (401), forbidden (403), not-found (404), conflict (409), payload-too-large (413), unprocessable (422), rate-limited (429), internal-error (500), unavailable (503). The shared route guard can also send principal-kind-unavailable (503). Other routes may omit it."
          },
          "needs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "description": "Common error fields where emitted; validation issues are present on some 400 responses. `error` is on every error body and `code` on every error body from the decision, workflow, agent and company-data routes; other fields are not guaranteed across all routes.",
        "additionalProperties": true
      },
...
      "PublishedWorkflowVersion": {
        "type": "object",
        "description": "One immutable, digest-verified workflow version and where it is in the publication lifecycle.",
        "required": ["workflowId", "version", "graph", "digest", "authoredBy", "createdAt", "status"],
        "properties": {
          "workflowId": { "type": "string" },
          "version": { "type": "integer", "minimum": 1 },
          "graph": { "$ref": "#/components/schemas/WorkflowGraph" },
          "digest": { "type": "string", "description": "Digest of the graph content; recomputed and checked on every read." },
          "authoredBy": { "type": "string" },
          "createdAt": { "type": "integer", "description": "Epoch milliseconds." },
          "status": { "type": "string", "enum": ["draft", "in-review", "published", "deprecated"] },
          "reviewedBy": { "type": "string" },
          "reviewNote": { "type": "string" },
          "reviewedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "publishedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "deprecatedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "rollbackOfVersion": { "type": "integer", "minimum": 1 }
        }
      },
...
      "WorkflowGraph": {
        "type": "object",
        "required": [
          "schemaVersion",
          "id",
          "version",
          "entryNodeId",
          "nodes",
          "edges"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "entryNodeId": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            }
          }
        },
        "additionalProperties": false
      },
      "WorkflowNode": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "label"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "trigger",
              "agent",
              "tool",
              "condition",
              "output"
            ]
          },
          "label": {
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/WorkflowNodeConfig"
          }
        },
        "additionalProperties": false
      },
      "WorkflowNodeConfig": {
        "type": "object",
        "properties": {
          "agentId": {
            "type": "string"
          },
          "toolId": {
            "type": "string"
          },
          "conditionExpression": {
            "type": "string"
          },
          "impact": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high",
              "critical"
            ]
          },
          "evaluationRequired": {
            "type": "boolean"
          },
          "sideEffect": {
            "type": "boolean"
          },
          "supervisorApprovalRequired": {
            "type": "boolean"
          },
          "maxAttempts": {
            "type": "integer"
          },
          "timeoutMs": {
            "type": "integer"
          }
        },
        "additionalProperties": false
      },
      "WorkflowEdge": {
        "type": "object",
        "required": [
          "id",
          "from",
          "to"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AgentManifest": {
        "type": "object",
        "required": [
          "id",
          "version",
          "authority",
          "tasks",
          "maximumImpact",
          "requiresEvaluation"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^nuera-quicksilver:[a-z][a-z0-9-]{0,62}$"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "authority": {
            "type": "string",
            "enum": [
              "propose",
              "review"
            ]
          },
          "tasks": {
            "type": "array",
            "minItems": 1,
            "maxItems": 10,
            "items": {
              "type": "string",
              "enum": [
                "reasoning",
                "code",
                "bulk",
                "planning",
                "routing",
                "tool",
                "memory",
                "hydraulic",
                "evaluation",
                "other"
              ]
            }
          },
          "maximumImpact": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high",
              "critical"
            ]
          },
          "requiresEvaluation": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AgentDefinitionInput": {
        "type": "object",
        "required": [
          "displayName",
          "description",
          "manifest"
        ],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          },
          "manifest": {
            "$ref": "#/components/schemas/AgentManifest"
          }
        },
        "additionalProperties": false
      },
      "AgentVersionAction": {
        "type": "object",
        "required": [
          "agentId",
          "version"
        ],
        "properties": {
          "agentId": {
            "$ref": "#/components/schemas/AgentId"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "ReviewAgentDefinition": {
        "type": "object",
        "required": [
          "agentId",
          "version",
          "note"
        ],
        "properties": {
          "agentId": {
            "$ref": "#/components/schemas/AgentId"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "note": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          }
        },
        "additionalProperties": false
      },
      "RollbackAgentDefinition": {
        "type": "object",
        "required": [
          "agentId",
          "sourceVersion"
        ],
        "properties": {
          "agentId": {
            "$ref": "#/components/schemas/AgentId"
          },
          "sourceVersion": {
            "type": "integer",
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "AgentId": {
        "type": "string",
        "pattern": "^nuera-quicksilver:[a-z][a-z0-9-]{0,62}$"
      },
      "AgentDefinition": {
        "type": "object",
        "required": [
          "agentId",
          "displayName",
          "description",
          "version",
          "manifest",
          "digest",
          "authoredBy",
          "createdAt",
          "lifecycle"
        ],
        "properties": {
          "agentId": {
            "$ref": "#/components/schemas/AgentId"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "manifest": {
            "$ref": "#/components/schemas/AgentManifest"
          },
          "digest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "authoredBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "draft",
              "in-review",
              "published",
              "archived"
            ]
          },
          "reviewedBy": {
            "type": "string"
          },
          "reviewNote": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "integer"
          },
          "publishedAt": {
            "type": "integer"
          },
          "rollbackFrom": {
            "type": "object",
            "required": [
              "version",
              "digest"
            ],
            "properties": {
              "version": {
                "type": "integer",
                "minimum": 1
              },
              "digest": {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              }
            },
            "additionalProperties": false
          },
          "builtIn": {
            "const": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "requestBodies": {
...
      "CreateWorkflowDraftRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflowDraftRequest" } } } },
      "WorkflowVersionAction": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkflowVersionAction" } } } },
      "ReviewWorkflowRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewWorkflowRequest" } } } },
...
      "CreateAgentDraft": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentDefinitionInput"
            }
          }
        }
      },
      "AgentVersionAction": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentVersionAction"
            }
          }
        }
      },
      "ReviewAgentDefinition": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ReviewAgentDefinition"
            }
          }
        }
      },
      "RollbackAgentDefinition": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RollbackAgentDefinition"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque principal or configured supervisor token",
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "RateLimitResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Rate limited by the current in-process token bucket.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        }
      },
...
      "PublishedWorkflowVersionResponse": { "description": "The workflow version after the lifecycle action (201 for a new draft, 200 otherwise).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublishedWorkflowVersion" } } } },
...
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Route-specific error; status and body vary by failure."
      },
...
      "AgentDefinitionResponse": {
        "description": "Created or updated agent definition returned by the lifecycle operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AgentDefinition"
            }
          }
        }
      },
...
    },
...
  }
}